Added background, logo test function. Modifed logo hover effect.
