+++ src/nestdoorapp/static/css/homepage_lu.css
@@ -1,0 +1,166 @@
+/* global style -- lu */
+* {
+  margin: 0;
+  padding: 0;
+}
+
+/* set the frame for the element -- lu */
+.wrapper {
+  position:absolute;
+  width: 100%;
+}
+
+/* specifiied attributes -- lu */
+.wrapper .image {
+  margin: 0;
+  padding: 0;
+  height: 100vh; /* Set the body's height to cover the entire viewport */
+  background: url('../images/reed_logo1.png');
+  background-size: 200px 400px;
+  background-position: center;
+  background-repeat: repeat; 
+  opacity: 0.2;
+  mix-blend-mode: multiply;
+}
+
+
+body {
+    background-color: floralwhite;
+    padding-top: 20px;
+  }
+  h1 {
+    color: black;
+  }
+  p {
+    color: white;
+  }
+
+
+
+/* Horizontal and Veritcal line Styling */  
+.h_line {
+    border-bottom: 6px solid darkgreen;
+    padding-top: 15px;
+    width: auto;
+    margin: auto;
+}
+
+
+.vl {
+    border-left: 6px solid darkgreen;
+    padding-left: 23%;
+    padding-top: 15%;
+    height: 400px;
+    margin-inline: 50%;
+}
+
+/* Text Spanning */
+.split-para { 
+  display:block;
+  margin:100px;
+  padding-left: 25%;
+}
+.split-para span { 
+  display:block;
+  float:right;
+  width:50%;
+  margin-left:100px;
+}
+
+/* Page set up wrapping */
+#page-container {
+  position: relative;
+  min-height: 100vh;
+}
+
+#content-wrap {
+  padding-bottom: 80px;    /* Footer height from above it */
+}
+
+/* Top navigation */
+  .topnav {
+    background-color: darkgreen;
+    overflow: hidden;
+    text-align: center;
+    position: fixed;
+    display: block;
+    transition: top 0.3s;
+    width: 100%;
+    top: -5px;
+  }
+  
+  /* Style the links inside the navigation bar */
+  .topnav a {
+    float: left;
+    color: floralwhite;
+    display: block;
+    text-align: center;
+    padding: 14px 16px;
+    text-decoration: none;
+    font-size: 25px;
+  }
+
+
+/* Apply hover effect to text links within .topnav, excluding the entire anchor tag -- lu */
+.topnav a:not(:has(img)):hover {
+  background-color: whitesmoke;
+  color: darkgreen;
+}
+
+
+  /* Add a color to the active/current link */
+  .topnav a.active {
+    background-color: rgb(100, 147, 100);
+    color: whitesmoke;
+  }
+
+  .center {
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+    width: 50%;
+  }
+
+
+/* Footer Styling */
+  .footer {
+    position: absolute;
+    bottom: 0;
+    width: 100%;
+    height: 90px;
+    background-color: darkgreen;
+    color: floralwhite;
+    text-align: center;
+  }
+
+  .footer-text-left {
+    font-size:25px;
+    padding-left:40px;
+    float:left;
+  }
+
+  .footer-text-right {
+    font-size:25px;
+    padding-right:40px;
+    float:right;
+  }
+
+  .footer-content-center {
+    padding-block-start: 0%;
+    margin-top:20px;
+    float:right;
+  }
+
+  .icon-style {
+    height:40px;
+    padding-right: 10px;
+    margin-left:0px;
+    margin-top:0px;
+  }
+
+
+  /* no change background when hovering on logo --lu */
+  .icon-style a:not(:has(img)):hover {
+    background-color:yellow;
+    padding:5px;
+  }
